[TEST] update CI trigger for offline test workflow (#521)
Update trigger for offline test workflow
Add test using scoop for Windows runner
Bugfixes to test scripts

+++ tests/scripts/cleanUpRunner.test.js
@@ -5,8 +5,8 @@ import fs from "fs";
 
 export function runCleanUp({
   id = 0,
-  installFolder,
-  toolsFolder,
+  installFolder = null,
+  toolsFolder = null,
   deleteAfterTest,
 }) {
   describe(`${id}- Clean up EIM folders from Runner |`, function () {
@@ -18,24 +18,30 @@ export function runCleanUp({
           `Running clean up function with Delete After Test = ${deleteAfterTest}`
         );
         try {
-          fs.rmSync(installFolder, { recursive: true, force: true });
-          logger.info(`Successfully deleted ${installFolder} folder`);
-          fs.rmSync(toolsFolder, {
-            recursive: true,
-            force: true,
-          });
-          logger.info(`Successfully deleted ${toolsFolder} folder`);
+          if (installFolder) {
+            fs.rmSync(installFolder, { recursive: true, force: true });
+            logger.info(`Successfully deleted ${installFolder} folder`);
+          }
+          if (toolsFolder) {  
+            fs.rmSync(toolsFolder, { recursive: true, force: true });
+            logger.info(`Successfully deleted ${toolsFolder} folder`);
+          }
         } catch (err) {
           logger.info(`Error deleting installation folders`);
+          logger.info(`Error: ${err}`);
         }
       }
     });
 
     it("1- Install and Tools folder should exist", async function () {
       logger.info("Validating folders exist before deleting");
-      expect(fs.existsSync(installFolder), "IDF Installation folder missing").to
-        .be.true;
-      expect(fs.existsSync(toolsFolder), "IDF tools folder missing").to.be.true;
+      if (installFolder) {
+        expect(fs.existsSync(installFolder), "IDF Installation folder missing").to
+          .be.true;
+      }
+      if (toolsFolder) {
+        expect(fs.existsSync(toolsFolder), "IDF tools folder missing").to.be.true;
+      }
     });
   });
 }
